requirements.txt
# requirements.txt
streamlit==1.32.0
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
dnspython==2.6.1